requirements.txt
attrs==19.3.0
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyspellchecker==0.5.4
pytest==5.4.3
regex==2020.6.8
six==1.15.0
wcwidth==0.2.5
zipp==3.1.0
